requirements.txt
alpaca-py
pandas
pytz
pyyaml
pyarrow
matplotlib
scipy
duckdb
torch
scikit-learn
seaborn
yfinance
pandas_ta
numpy
xgboost
tensorflow